update dcm config and rules
