+++ analysis_options.yaml
@@ -11,14 +11,21 @@ analyzer:
     - dart_code_metrics
 
 dart_code_metrics:
+  extends:
+    - package:dart_code_metrics_presets/all.yaml
   metrics:
     cyclomatic-complexity: 20
     number-of-parameters: 4
     maximum-nesting-level: 5
+    source-lines-of-code: 500
   metrics-exclude:
     - test/**
     - integration_test/**
   rules:
+    - newline-before-return
+    - no-boolean-literal-compare
+    - no-empty-block
+    - prefer-trailing-comma
     - avoid-dynamic
     - avoid-redundant-async
     - avoid-unnecessary-type-assertions
@@ -31,11 +38,6 @@ dart_code_metrics:
     - no-equal-then-else
     - avoid-nested-conditional-expressions:
       acceptable-level: 2
-    - prefer-trailing-comma
-    # - avoid-passing-async-when-sync-expected
-    # - newline-before-return
-    # - prefer-moving-to-variable
-    # - prefer-match-file-name
 
 # The following line activates a set of recommended lints for Flutter apps,
 # packages, and plugins designed to encourage good coding practices.
@@ -54,7 +56,7 @@ linter:
   # `// ignore_for_file: name_of_lint` syntax on the line or in the file
   # producing the lint.
   rules:
-    # avoid_print: false  # Uncomment to disable the `avoid_print` rule
+    avoid_print: false  # Uncomment to disable the `avoid_print` rule
     # prefer_single_quotes: true  # Uncomment to enable the `prefer_single_quotes` rule
 
 # Additional information about this file can be found at
