+++ lib/router/routes.dart
@@ -1,0 +1,8 @@
+class ScreenPaths {
+  static String root = '/';
+  static String splash = '/splash';
+  static String home = '/home';
+  static String screen1 = '/screen1';
+  static String screen2 = '/screen2';
+  static String settings = '/settings';
+}
