+++ pubspec.yaml
@@ -51,13 +51,14 @@ dev_dependencies:
   flutter_test:
     sdk: flutter
 
-  flutter_lints: ^4.0.0
   json_serializable: ^6.7.1
   mocktail: ^1.0.4
-  bloc_test: ^9.1.7
-  analyzer_plugin: ^0.11.3
-  
   mockito: ^5.4.3
+  bloc_test: ^9.1.7
+  
+  flutter_lints: ^4.0.0
+  analyzer_plugin: ^0.11.3
+  dart_code_metrics_presets: ^2.16.0
 
 flutter:
   uses-material-design: true
